Time: 4 ms (21.38%), Space: 8.2 MB (10.06%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     double myPow(double x, int n) {
       double ans = 1.0;
-        long long m = abs(n); // Handle the edge case of n being INT_MIN
+        int m = abs(n); // Handle the edge case of n being INT_MIN
 
         while (m > 0) {
             if (m % 2 == 1) {
